refactor(mcp): distribute server.rs tests into per-module tests.rs (lab-kvji.24.1.6)
Move the ~866 LOC test block out of server.rs into <module>/tests.rs child files
(no-mod.rs child-module pattern, preserving super:: private access):
- result_format/tests.rs: estimate_tokens*, extract_error_info*, canonical_kind, tool_error_envelope
- upstream/tests.rs: normalize_upstream_result
- call_tool_codemode/tests.rs: code_mode_filter_arg_*, code_execute_description, gateway_search_schema
- completion/tests.rs: completion_* + completion_test_registry fixture
- handlers_tools/tests.rs: snapshot_catalog_*, server_reads_current_pool, service_actions_json (#[ignore])
- context/tests.rs: scope/admin/oauth tests + make_auth fixture
- server.rs keeps only server_capabilities + upstream_subject_resolution_self_test

Fixtures (completion_test_registry/noop_dispatch) duplicated where needed (<40 LOC).
Leaf test names IDENTICAL to baseline, total count 1447 unchanged — only module paths shifted.
server.rs 1196 -> 373 LOC. Every tests.rs <500 (max 286). 1423 tests pass.
