+++ crates/lab/src/mcp/call_tool_codemode.rs
@@ -373,3 +373,6 @@ impl LabMcpServer {
         Ok(CallToolResult::success(vec![Content::text(output)]))
     }
 }
+
+#[cfg(test)]
+mod tests;
